Allow either $(Configuration) or $(WixFlavor) to select debug/ship build type.
The WixFlavor property is supposed to be internal to the build system, but if
specified on the command-line can lead to debug binaries being put in the ship
directory. Fix this by allowing either to control whether a debug or ship build
is done.  Since $(Configuration) is preferred, update the documentation to
mention that instead of $(WixFlavor).

--HG--
branch : wix38
extra : rebase_source : d369c85d96e2fc2d4421a65e8cf6ecaba14cbf3d