Handle MoveFileEx unexpectedly returning ERROR_FILE_NOT_FOUND.
We are seeing some cases where ::MoveFileEx() says a file was not found
but the source file is actually present. In that case, return path not
found so we try to create the target path since that is most likely
what is missing. Otherwise, the source file is missing and we're obviously
not going to be recovering from that.