Fix #4083 by providing a table of contents at the root of the online manual.
This change enhances DocCompiler to be able to append a TOC in Markdown
on to an existing file in the the documentation (in our case, the root
index.html.md). It also adds the ability to ignore files from the CHM
(so the TOC is not included in the CHM) and define a the default document
in the CHM (rather than always make it the root).