Move wix.targets to Microsoft.Common.targets.
This is a major change to utilize the standard Microsoft.Common.targets for
wix.targets This should allow us to integrate with other projects better
than we have in the past. It also allows us to remove significant amounts
of MSBuild targets that duplicated what was in Microsoft.Common.targets.