High probability fix #3854 to address memory corruption on subsequent prints.
This bug does not reproduce consistently so it was very challenging to
hunt down. A previous change started by cleaning up memory that wasn't
correctly freed but it wasn't clear that would actually fix the root
issue. This change nulls out a global string pointer that on subsequent
executions would previously be left pointing to released memory and
could incorrectly be accessed. It is far more likely this is the root
cause of the issue.