Issue 4140 was caused by burn writing itself into the "RunOnce" registry-path to make sure it is being rerun after a mid-install reboot.
Unfortunately on Windows XP, a driver install during one of the chained MSI-packages would cause the RunOnce entries to be
executed prematurely.  This caused the startup-screen to be shown again next to the "Success" screen of the complete installation-
procedure.  For operating-systems prior to Vista now the "Run" registry-path is used instead of "RunOnce".